Add MidnightBSD and fix a few other things
1. Add MidnightBSD to the operating system list.
Wikipedia Link: https://en.wikipedia.org/wiki/MidnightBSD

2. Fix FreeBSD porters handbook URL. The previous URL
leads to a non-existent webpage.

3. __FreeBSD_version shall be included from sys/param.h
and not from osreldate.h. It looks like osreldate.h exists
for compatibility reasons as __FreeBSD_version is defined
0 in there and including sys/param will undefine the
__FreeBSD_version macro anyway that's defined in osreldate.h.

4. Fix the Wikipedia links, (e.g. use BSD, (uppercase)
and not bsd (lowercase)).

5. Change the "notice" for Android, to make it more clear.
Android just uses Linux as the kernel but it isn't properly
"based" on it.