fix links, add more references, and indent the source
Standards.md:
1. Add C89, and C90 references, and point them to Wikipedia.

2. C99 and C11, use HTTPS instead of the HTTP URL.

3. Use C17 term, and not C18. Although both are just valid, the
widely used term is C17, and C18 can be confusing (for those who
don't know that both are the same).

4. C17, C++{98,11,14,17}, use HTTPS.

5. The original document for C++/CLI is no longer available, and I
can only able to find an archived version (including the PDF) on
the Internet Archive. So replace the original URL with the archived
one.

6. DSP-C is also no longer available. Fortunately, there's
an archived version of the page, completely functional. Replace
the original URL with the archived one.

7. Add PDFs for POSIX.1-1988 and POSIX.1-1990.

8. SUSv{2,3,4} use HTTPS.

Endianness.md:
1. CPU byte order convertion shouldn't be done using htons() function.
htons() is defined and used to convert host to network byte order, and
if the host and network server both are consists different endians,
this will not work, as if the host is big-endian, htons() just returns
the input value and no convertion is performed. Also, considering,
htons() only accept unsigned short, it also need to conver htonl()
(unsigned int).

2. Indent the example code.

DataModels.md:
Use HTTPS, as HTTP request is unreachable (blocked by their firewall).